dictionary: Respond to our nick
When someone says the bot's nickname it will output a random chatty
response. Previously it did not respond at all. Note this will only work
for cases where the bot is not directly addressed, as we expect those
cases to indicate a command to the bot.

This also resolves a potential bug where string match metacharacters
were unescaped when looking for terms in messages.